Video Analyzer changes

+++ Data_analysis/CodeProfiler.py
@@ -21,7 +21,7 @@ def NewFrame():
     FrameCounter += 1
     period = time.time() - FrameAveragePeriodStart
     if period > 20:
-        framerate = int(FrameCounter/period)
+        framerate = FrameCounter/period
         print(f'Frame Rate= : {framerate:.2f}')
         FrameAveragePeriodStart = time.time()
         FrameCounter = 0
